+++ CHANGELOG.md
@@ -12,7 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Negative reference temperatures for the fluid are now possible (issue #192).
 
 ## fixed
-- Bug in load-duration curve when not working with optimize load profile (issue # 189).
+- Bug in load-duration curve when not working with optimize load profile (issue #189).
+- Bug in hourly data (issue #196).
+- Bug in saving after a file has been moved (issue #198).
 
 
 ## [2.2.0] - 2023-10-17